Add support for serialising/desieralising RegexLexer.
Having lexers defined only in Go is not ideal for a couple of reasons.
Firstly, it impedes a significant portion of contributors who use Chroma
in Hugo, but don't know Go. Secondly, it bloats the binary size of any
project that imports Chroma.

Why XML? YAML is an abomination and JSON is not human editable. XML
also compresses very well (eg. Go template lexer XML compresses from
3239 bytes to 718).

Why a new syntax format? All major existing formats rely on the
Oniguruma regex engine, which is extremely complex and for which there
is no Go port.

Why not earlier? Prior to the existence of fs.FS this was not a viable
option.

Benchmarks:

    $ hyperfine --warmup 3 \
        './lexers.test-pre-xml -test.count=1' \
        './lexers.test-xml -test.count=1' \
        './lexers.test-xml-xz -test.count=1' \
        './lexers.test-xml-gz -test.count=1'
    Benchmark 1: ./lexers.test-pre-xml -test.count=1
      Time (mean ± σ):     383.4 ms ±   2.0 ms    [User: 442.1 ms, System: 19.2 ms]
      Range (min … max):   381.1 ms … 388.3 ms    10 runs

    Benchmark 2: ./lexers.test-xml -test.count=1
      Time (mean ± σ):     421.8 ms ±   1.4 ms    [User: 490.1 ms, System: 19.7 ms]
      Range (min … max):   419.6 ms … 424.4 ms    10 runs

    Benchmark 3: ./lexers.test-xml-xz -test.count=1
      Time (mean ± σ):     540.1 ms ±   3.4 ms    [User: 891.2 ms, System: 39.0 ms]
      Range (min … max):   536.1 ms … 547.5 ms    10 runs

    Benchmark 4: ./lexers.test-xml-gz -test.count=1
      Time (mean ± σ):     428.2 ms ±   1.9 ms    [User: 503.4 ms, System: 19.0 ms]
      Range (min … max):   425.7 ms … 431.9 ms    10 runs

    Summary
      './lexers.test-pre-xml -test.count=1' ran
        1.10 ± 0.01 times faster than './lexers.test-xml -test.count=1'
        1.12 ± 0.01 times faster than './lexers.test-xml-gz -test.count=1'
        1.41 ± 0.01 times faster than './lexers.test-xml-xz -test.count=1'

And binary size difference:

    $ du -h lexers.test*
    9840	lexers.test-pre-xml
    8952	lexers.test-xml
    7949	lexers.test-xml-gz
    8204	lexers.test-xml-xz

XML without compression adds a 10% initialisation overhead and a 10%
size reduction. GZipped XML files adds 12% initialisation overhead but a
20% size reduction.